refactor: JdbcTemplate 통합
